+++ src/services/__mocks__/data/data-utils.js
@@ -1,0 +1,2 @@
+export const makeArrayOfMocks = (mockFn, count) =>
+  [...Array(count).keys()].map((e) => mockFn())
